Move core implementation to chunked buffer
